update code at home screen
