+++ lib/res/dimentions/gaps.dart
@@ -1,0 +1,4 @@
+import 'package:flutter/material.dart';
+
+const gapH1 = SizedBox(height: 10);
+const gapH2 = SizedBox(height: 20);
