+++ lib/screens/home_view/components/authentication.dart
